add fetch frozen reports & download reports functionality
