+++ package.json
@@ -24,7 +24,10 @@
     "@hookform/resolvers": "^3.3.1",
     "classnames": "^2.3.2",
     "formik": "^2.4.6",
+    "html2canvas": "^1.4.1",
+    "jspdf": "^2.5.1",
     "moment": "^2.30.1",
+    "pdf-lib": "^1.17.1",
     "react-hook-form": "^7.46.2",
     "swr": "^2.2.5",
     "xlsx": "^0.18.5",
