+++ packages/esm-report-app/src/location/location.scss
@@ -1,6 +1,3 @@
-@use '@carbon/colors';
-@use '@carbon/layout';
-
-.loader {
-  margin: 2rem;
+.container {
+  margin: 16px 0;
 }
