--- packages/esm-report-app/src/registers/MOH-362/MOH_362_component.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import RenderMOH362a from './MOH-362a';
-import RenderMOH362b from './MOH-362b';
-
-const RenderMOH362: React.FC = () => (
-  <div>
-    <RenderMOH362a />
-    <RenderMOH362b />
-  </div>
-);
-export default RenderMOH362;
